Move Assay search categories
